Fix debug info cloning when rewriting varargs functions.
Debug info should be unique, and thus pointers to debug metadata are held
outside of the PGLTEntryWrappers pass. Cloning a function results in incorrect
debug info for inlined functions, since the metadata is cloned into new
objects. We don't actually need to clone the function to replace it, so the pass
now moves function contents directly into the replacement function, leaving all
metadata as is.