Move PGLT bin mapping to AsmPrinter
Functions are now assigned to bins simply by adding a prefix to their section
name. Since sections are already uniqued by the MCContext, we know that
functions with the same section prefix will end up in the same
section/bin (assuming all other things are equal).

The AsmPrinter now lazily constructs the PGLT when PGLT entries are needed by
adding sections (bins) to an array of sections as they are referenced. When the
PGLT is emitted, the AsmPrinter emits section beginning symbols according to the
ordering in the PGLT section array.

We only ever need to resolve a PGLT offset or bin offset during the AsmPrinting
phase, so it makes sense to keep the bin referencing API exclusively in
AsmPrinter.

Change-Id: Iac2f88b8a1db61de0aef3771d52943896e5afecd