Move PGLT bin mapping to AsmPrinter (#6)
* Move PGLT bin mapping to AsmPrinter

Functions are now assigned to bins simply by adding a prefix to their section
name. Since sections are already uniqued by the MCContext, we know that
functions with the same section prefix will end up in the same
section/bin (assuming all other things are equal).

The AsmPrinter now lazily constructs the PGLT when PGLT entries are needed by
adding sections (bins) to an array of sections as they are referenced. When the
PGLT is emitted, the AsmPrinter emits section beginning symbols according to the
ordering in the PGLT section array.

We only ever need to resolve a PGLT offset or bin offset during the AsmPrinting
phase, so it makes sense to keep the bin referencing API exclusively in
AsmPrinter.

Change-Id: Iac2f88b8a1db61de0aef3771d52943896e5afecd

* Simplify ARMPGLTOpt::isSameBin.

Change-Id: I5d38c978579809a848c00161b54effda23c98eb5

* Make AsmPrinter::PGLT private since it can be

Change-Id: I12815df7130542fa041f9bc0da5a7fdd51fffcc5

* Reduce code paths/complexity and add comment for default bin #0

Change-Id: I53f2f2988c0064cfd005fd4a28e04d072fac84bc

* Add comment for call to Function::setSectionPrefix

Change-Id: Ifd6b81b602388c506ed89ca0dade25bebc6e2f6b