Make naming of wrapper and original function consistent
If a wrapper is created it will now always take the name of the
original function even if not strictly necesary (i.e., for
functions which only need a wrapper because they have their
address taken).
The original function will be renamed to <name>$$orig
(<name>$$origva for vararg functions).