[pagerando] Add target-independent backend modifications for pagerando
This patch adds several features to support codegen for pagerando. We add a new
ISD opcode and pseudo-source value for the page offset table (analogous to the
global offset table) which allows backends to emit references to the POT. The
patch extends target-independent lowering slightly to store a POT base register
for backends that support pagerando, and handles interactions between pagerando
sections, unique sections, and jump tables. Finally, the bulk of the patch
supports emission of the POT in AsmPrinter.

Differential Revision: https://reviews.llvm.org/D37584

Change-Id: Ie9b3b6064101db1cf241f6ac896e054f151a75e2