Add Gold plugin LTO support for pagerando
Pagerando bins need to be assigned to unique segments and marked with a flag to
designate the segment as pagerando-compatible. This patch adds support to the
gold plugin to assign pagerando bin sections to unique PF_RAND_ADDR flagged
segments.

This patch depends on gold to allow plugins to assign sections from plugin-generated
files to unique segments. This feature was added to gold in commit
c4e648430f3c5c135310f87280d45dba581aaa7b on 2017-12-11.

Differential Revision: https://reviews.llvm.org/D37587

Change-Id: I632d2fcd45866417a700a60123eb742243ebe017