Pass profdata file, if present, to compiler build
Bug: http://b/37574207

The profdata file is temporarily planned to be checked into
llvm/android/profiles.  We can move it to a separate project if
necessary.

The profile file will be named <clang-long_version>.profdata.  The
compiler build will use this profile if it exists.  The build will
continue if this file doesn't exist, unless --check-pgo-profile is
specified.  (Just one of the builds in go/ab/aosp-llvm-toolchain will
set this flag to catch missing profiles).

A new flag, --no-pgo, is added to test_compiler.py to skip using the
profile file even if it is present.  A build in
go/ab/aosp-master-plus-llvm will use this to track PGO vs. Non-PGO
speedup.

Test: With profdata from a local build, run
          - build.py with and without --check-pgo-profile
          - test_compiler.py with and without --no-pgo

Change-Id: Ifdca75c479197580d26c1297fe672d17622ddfb5