Ignore function uses in hidden vtables when determining wrappers
Hidden/internal vtables are created for classes marked with hidden
visibility. Hidden classes should never be called externally since the external
user will not be able to call any non-virtual methods of the class. Thus, we can
safely skip wrappers for local methods in hidden classes, even if they are
address-taken for a hidden vtable.

Change-Id: Ib975338c169addfa27be9d969bd73f41f60b6946