Remove install directories at the start
Bug: http://b/28196329

Remove install directories at the start, so we get a clean slate when
changing install dirs etc. or updating any of the custom copies we do
for the runtimes.

Don't clean stage1-install to better support incremental builds.  This
is fine since there is just one top level Clang build that installs
there.

Test: python toolchain/llvm/android/build.py
Change-Id: I95d1118de59da5da38ed5efa4ccff78ce54f6c87