Remove pagerando_wrapper attribute and rename pagerando_binned
We are not actually using the pagerando_wrapper function attribute for anything
at the moment, so for the sake of simplicity and not unnecessarily adding to the
external attribute API, we can remove that attribute. In the future we may need
to add this attribute back to identify and specially process pagerando wrapper
functions.

Since pagerando_wrapper is gone, pagerando_binned is now renamed to be just
pagerando for further simplicity.

Adds docs to LangRef for the pagerando attribute.

Change-Id: Id57ffa7edc4c5b084387bcaf37e0803da1b5052a