Build Fuzzer and OpenMP against NDK and platform libc++
Bug: http://b/28196329

The NDK libc++ puts symbols in a separate namespace and are not
available for platform libraries.  Build separate versions of Clang
runtimes that use C++ standard library (libFuzzer, libomp) - one against
the C++ headers (which are the same as in
<clang-prebuilts>/include/c++/v1) and one against the C++ from the NDK.

The prebuilts against the platform are also built at API level 26.

Test: python toolchain/android/llvm/build.py and build a platform
library that uses the libFuzzer from the new prebuilts.

Change-Id: Ic0811c3ab5433ee9296e22ce16de9b8a328f8f52