[MemorySSA] Be less aggressive with @llvm.lifetime.start
This is a workaround to keep LLVM from incorrectly removing stores in
rare cases. It has not yet been committed, and -- due to the direction
we seem to want to take MSSA in -- probably never will be.

It should be sufficient to work around the problem until we agree on a
good way forward.

Additional context: https://reviews.llvm.org/D43269

This is rebased to r327695 by gbiv@chromium.org.

Change-Id: Ib78a9e12845a28a48bbc4154e89775b1ef4b5319